+++ covasim/sim.py
@@ -289,13 +289,13 @@ class Sim(cvbase.BaseSim):
         # nonzero_betas = betas[nonzero_inds]
         # nonzero_targets = targets[nonzero_inds]
 
-        beta = np.float32(self['beta'])
-        transmission_inds = cvu.compute_targets(beta, np.array(sources, dtype=np.int32), np.array(targets, dtype=np.int32), layer_betas, people.rel_trans, people.rel_sus)
-        # transmission_inds = np.array([], dtype=np.int32)
+        beta = np.float64(self['beta'])
+        transmission_inds = cvu.compute_targets(beta, np.array(sources, dtype=np.int64), np.array(targets, dtype=np.int64), layer_betas, people.rel_trans, people.rel_sus)
+        # transmission_inds = np.array([], dtype=np.int64)
 
         # Calculate actual transmission
 
-        # transmission_inds = cvu.bernoulli_filter(np.float32(self['beta']), np.int32(targets))
+        # transmission_inds = cvu.bernoulli_filter(np.float64(self['beta']), np.int64(targets))
         flow_counts['new_infections'] += people.infect(inds=transmission_inds, t=t)
 
         # Apply interventions
